+++ test/noahtheduke/splint/pattern_test.clj
@@ -4,14 +4,13 @@
 
 (ns noahtheduke.splint.pattern-test
   (:require
-   [clojure.test :refer [deftest]]
-   [expectations.clojure.test :refer [expect from-each expecting]]
+   [expectations.clojure.test :refer [defexpect expect from-each expecting]]
    [noahtheduke.splint.pattern :as sut]
    [noahtheduke.splint.test-helpers :refer [parse-string]]))
 
 (set! *warn-on-reflection* true)
 
-(deftest read-dispatch-test
+(defexpect read-dispatch-test
   (expecting "simple types"
     (doseq [[input t] '[[nil :nil]
                         [true :boolean]
@@ -32,7 +31,7 @@
       (expect t (sut/read-dispatch input))
       (expect :list (sut/read-dispatch (vary-meta input assoc :splint/lit true))))))
 
-(deftest match-any-test
+(defexpect match-any-test
   (expect {}
     (from-each [input '[a 'a :a "a" 1 true nil
                         (1 2 3)
@@ -41,7 +40,7 @@
                         [1 2 3]]]
       ((sut/pattern '(1 2 _)) (list 1 2 input)))))
 
-(deftest predicate-test
+(defexpect predicate-test
   (expect {}
     ((sut/pattern '(1 2 (? _ symbol?))) '(1 2 a)))
   (expect nil?
@@ -54,7 +53,7 @@
   (expect '{?a a}
     ((sut/pattern '(1 2 (? a symbol?))) '(1 2 a))))
 
-(deftest ?*-test
+(defexpect ?*-test
   (expect
     '{?exprs []}
     ((sut/pattern '[(?* exprs)])
@@ -80,7 +79,7 @@
     ((sut/pattern '((?* exprs) 1 (?* exprs)))
      (parse-string "(1 2 3 1 2 3)"))))
 
-(deftest ?+-test
+(defexpect ?+-test
   (expect
     nil?
     ((sut/pattern '[(?+ exprs)])
@@ -106,7 +105,7 @@
     ((sut/pattern '((?+ exprs) 1 (?+ exprs)))
      (parse-string "(1 2 3 1 2 3)"))))
 
-(deftest ??-test
+(defexpect ??-test
   (expect
     '{?exprs []}
     ((sut/pattern '[(?? exprs)])
@@ -145,7 +144,7 @@
     ((sut/pattern '((?? exprs)))
      (parse-string "(1 2 3)"))))
 
-(deftest ?|-test
+(defexpect ?|-test
   (expect
     nil?
     ((sut/pattern '[(?| exprs [1 2])])
@@ -179,12 +178,12 @@
     ((sut/pattern '[(?| exprs [1 2]) 3])
      (parse-string "[2 3]"))))
 
-(deftest quote-in-pattern-test
+(defexpect quote-in-pattern-test
   '{}
   ((sut/pattern '(a b 'c))
    (parse-string "(a b 'c)")))
 
-(deftest literals-test
+(defexpect literals-test
   ;; [[sut/pattern]] calls drop-quote so have to double up on
   ;; quoted form
   (expect (sut/pattern 'a) (parse-string "a"))
@@ -195,31 +194,31 @@
   (expect (sut/pattern 'true) (parse-string "true"))
   (expect (sut/pattern 'nil) (parse-string "nil")))
 
-(deftest list-test
+(defexpect list-test
   (let [pat (sut/pattern '(:a 1 :b [2] :c {:d 3}))]
     (expect not (pat (parse-string "(:a 1 :b [2] :c {:e 4})")))
     (expect (pat (parse-string "(:a 1 :b [2] :c {:d 3})")))
     (expect not (pat (parse-string "(:a 1 :b [2] :c {:d 3} :e 4)")))))
 
-#_(deftest map-test
+#_(defexpect map-test
     (let [pat (sut/pattern '{:a 1 :b [2] :c {:d 3}})]
       (expect not (pat (parse-string "{:a 1 :b [2] :c {:e 4}}")))
       (expect (pat (parse-string "{:a 1 :b [2] :c {:d 3}}")))
       (expect (pat (parse-string "{:a 1 :b [2] :c {:d 3} :e 4}")))))
 
-#_(deftest set-test
+#_(defexpect set-test
     (let [pat (sut/pattern '#{:a 1 :b [2] :c {:d 3}})]
       (expect not (pat (parse-string "#{:a 1 :b [2] :c}")))
       (expect (pat (parse-string "#{:a 1 :b [2] :c {:d 3}}")))
       (expect (pat (parse-string "#{:a 1 :b [2] :c {:d 3} :e [4]}")))))
 
-(deftest vector-test
+(defexpect vector-test
   (let [pat (sut/pattern '[:a 1 :b [2] :c {:d 3}])]
     (expect not (pat (parse-string "[:a 1 :b [2] :c {:e 4}]")))
     (expect (pat (parse-string "[:a 1 :b [2] :c {:d 3}]")))
     (expect not (pat (parse-string "[:a 1 :b [2] :c {:d 3} :e 4]")))))
 
-(deftest expand-specials-test
+(defexpect expand-specials-test
   (expect 'a (sut/expand-specials 'a))
   (expect '(a b c) (sut/expand-specials '(a b c)))
   (expect '(a _ c) (sut/expand-specials '(a _ c)))