+++ test/noahtheduke/splint/rules/naming/predicate_test.clj
@@ -4,14 +4,14 @@
 
 (ns noahtheduke.splint.rules.naming.predicate-test
   (:require
-   [clojure.test :refer [deftest]]
+   [expectations.clojure.test :refer [defexpect]]
    [noahtheduke.splint.test-helpers :refer [expect-match single-rule-config]]))
 
 (set! *warn-on-reflection* true)
 
 (defn config [] (single-rule-config 'naming/predicate))
 
-(deftest predicate-test
+(defexpect predicate-test
   (expect-match
     [{:form '(defn is-palindrome [a] true)
       :alt '(defn palindrome? [a] true)}]